% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_vars.R
\name{correlate_vars}
\alias{correlate_vars}
\title{Generates variables correlated with the input}
\source{
Adapted from \url{https://stat.ethz.ch/pipermail/r-help/2007-April/128925.html}
}
\usage{
correlate_vars(x, cors, mean = mean(x), sd = sd(x))
}
\arguments{
\item{x}{the input variable to be correlated against}

\item{cors}{a vector of correlations to generate variables for}

\item{mean}{a vector of means for the correlated variables, defaulting to the
mean of the input}

\item{sd}{a vector of standard deviations for the correlated variables,
defaulting to the standard deviation of the input}
}
\value{
a data frame, whose first column is the input, and whose remaining
are correlated variables for each provided correlation
}
\description{
Given a (numeric) input variable, this function generates a set of variables
correlated to the input, with provided correlations. Currently it only
supports specification of correlations with the original input, not a full
covariance matrix among the generated variables. Depending on the provided
correlations and the random numbers generated, this will occasionally fail
for now due to an internal matrix not being positive definite. In that case,
you should be able to specify one variable at a time and bind the outputs.
}
