% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_colors}
\alias{ggplot_colors}
\title{Shows the default ggplot colors}
\usage{
ggplot_colors(n)
}
\arguments{
\item{n}{The number of colors needed}
}
\value{
A character vector of the hex codes, and a plot showing their colors
}
\description{
Shows the default ggplot colors
}
