% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negations.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{the object being tested}

\item{y}{the vector `x` might be in}
}
\value{
a logical vector
}
\description{
Negates \code{\%in\%}.
}
