% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_batch}
\alias{read_batch}
\title{Reads multiple files into one data frame}
\usage{
read_batch(dir = NULL, ...)
}
\arguments{
\item{dir}{The folder (or zip file) where the files should be read from}

\item{...}{Arguments to pass to \code{data.table::fread}}
}
\value{
A data frame containing all of data, with a column to identify which
file a row came from
}
\description{
This is essentially a wrapper for \code{data.table::fread}, which will read
a batch of files in the same location (directory or zip file).
}
