% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_analysis.R
\name{weighted_crosstab}
\alias{weighted_crosstab}
\title{A weighted two-way crosstab generator}
\usage{
weighted_crosstab(
  data,
  var1,
  var2,
  show_na = FALSE,
  drop_set = c("No response", "Not applicable", "Refuse"),
  output = c("n", "pct", "both"),
  totals = c("none", "as row", "as col", "both"),
  pct_denom = c("row", "col", "all"),
  round_n = 1,
  round_pct = 1
)
}
\arguments{
\item{data}{A weighted data object of class `svy_table`}

\item{var1}{The variable to be tabulated, or the grouping variable}

\item{var2}{(Optional) the variable to be tabulated within each `var1` group}

\item{show_na}{Whether rows/columns that are `NA` should be included}

\item{drop_set}{Which non-NA classes should be dropped}

\item{output}{Whether the cells should show weighted counts, proportions, or both}

\item{totals}{Whether totals should be included as a row, column, both, or neither}

\item{pct_denom}{Which direction the percentages should be calculated on}

\item{round_n}{How many digits to round counts to}

\item{round_pct}{How many digits to round percentages to}
}
\description{
Uses `tabyl.svy_table` to create and format a two-way crosstab.
}
