% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_analysis.R
\name{weighted_tabyl}
\alias{weighted_tabyl}
\title{Survey weighted method for `tabyl`}
\usage{
weighted_tabyl(data, var1, var2, show_na = TRUE, ...)
}
\arguments{
\item{data}{A weighted data object of class `svy_table`}

\item{var1}{The variable to be tabulated, or the grouping variable}

\item{var2}{(Optional) the variable to be tabulated within each `var1` group}

\item{show_na}{Whether rows/columns that are `NA` should be included}

\item{...}{Currently unused}
}
\value{
A `data.frame` containing weighted counts in each one-way or two-way
combination of the input variables.
}
\description{
`janitor::tabyl()`, while great, doesn't work with weighted data. This is a
start to fixing that; it takes a `svy_table` object and does `tabyl` on it
(roughly). Currently it only works for 1 or 2 variables, and really does not
look like `tabyl`'s internals so be careful.
}
