% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{writer}
\alias{writer}
\alias{chunk_wrapper}
\alias{secret_chunk_wrapper}
\title{Write lines of text to a file}
\usage{
writer(x, ..., outfile = NULL, sep = "\\n")

chunk_wrapper(
  x,
  ...,
  outfile = NULL,
  options = c("echo=FALSE", "warning=FALSE"),
  label = NULL
)

secret_chunk_wrapper(
  x,
  ...,
  outfile = NULL,
  options = c("echo=FALSE", "include=FALSE", "warning=FALSE", "message=FALSE",
    "error=FALSE"),
  label = NULL
)
}
\arguments{
\item{x}{The first piece of text to write}

\item{...}{Further text to write}

\item{outfile}{The file to write to}

\item{sep}{Character to separate arguments of \code{...}, defaults to new line}

\item{options}{Character vector detailing the R chunk options}

\item{label}{A name to give the R chunk}
}
\description{
Write lines of text to a file
}
\section{Functions}{
\itemize{
\item \code{chunk_wrapper}: Write an R chunk to a file

This is a wrapper for \code{writer} that formats the text as an R chunk

\item \code{secret_chunk_wrapper}: Write a secret R chunk to a file

Wrapper for \code{chunk_wrapper} with specific options set to hide output
}}

